Data types refactored
