+++ src/common/datamodeltypes/canrawsenderdata.h
@@ -1,0 +1,8 @@
+#ifndef CANRAWSENDERDATA_H
+#define CANRAWSENDERDATA_H
+
+#include "candevicedata.h"
+
+typedef CanDeviceDataIn CanRawSenderDataOut;
+
+#endif // CANRAWSENDERDATA_H
