+++ src/common/datamodeltypes/canrawviewdata.h
@@ -1,0 +1,8 @@
+#ifndef RAWVIEWDATA_H
+#define RAWVIEWDATA_H
+
+#include "candevicedata.h"
+
+typedef CanDeviceDataOut CanRawViewDataIn;
+
+#endif // RAWVIEWDATA_H
